Update the save_visualization function.
It can be seen from the code that the predicted_mask is saved in the form of a heat map. Your team adopted cv2.COLORMAP_HOT.  But after np.hstack, result[:, :, ::-1] is added. This should be used in the processing of the input image, not in the predicted_mask. I made some adjustments.